requirements.txt
PyYAML==3.13
paramiko==2.4.2
pycrypto==2.6.1
pexpect==4.6.0
requests==2.22.0